pyproject.toml
[project]
name = "lra"
version = "0.1.0"
description = "Learning Content Compliance Intelligence System — RAG-based DQC validation"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # ── Core LangChain ──
    "langchain-core>=0.3.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.1.0",
    "langchain-chroma>=0.2.0",

    # ── LLM Providers (optional, uncomment when needed) ──
    # "langchain-openai>=0.3.0",
    # "langchain-anthropic>=0.3.0",
    # "langchain-community>=0.3.0",

    # ── Vector DB ──
    "chromadb>=0.5.0",

    # ── Document Extraction ──
    "PyMuPDF>=1.24.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "openpyxl>=3.1.0",

    # ── API + Frontend ──
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "streamlit>=1.40.0",
    "python-multipart>=0.0.12",
    "websockets>=13.0",

    # ── Reporting ──
    "reportlab>=4.2.0",

    # ── Data / Config ──
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",

    # ── Logging ──
    "structlog>=24.4.0",

    # ── Utilities ──
    "tiktoken>=0.8.0",
    "numpy>=1.26.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "black>=24.10.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
